long

+++ AtCoder/Program.cs
@@ -17,7 +17,7 @@ namespace AtCoder
 	{
 		public static void Exec()
 		{
-			long n = int.Parse(Console.ReadLine());
+			long n = long.Parse(Console.ReadLine());
 
 			//Console.WriteLine($"{answer}");
